refactor(core): extract remote, geometry_detection, file_utils from common.py
Phase 1 of common.py split (Issue #363):
- core/remote.py: S3/GCS/Azure URL handling functions
- core/geometry_detection.py: Geometry column detection functions
- core/file_utils.py: Path resolution, glob patterns, caching utilities

All functions are re-exported from common.py for backwards compatibility.
Uses lazy imports in extracted modules to avoid circular dependencies.

Reduces common.py from 4,161 to 3,352 lines (19.4% reduction).
